refactor: Amount를 dto 패키지로 이동
