+++ src/main/java/domain/LottoService.java
@@ -1,5 +1,7 @@
 package domain;
 
+import dto.Amount;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
