+++ src/main/java/view/LottoInputView.java
@@ -25,6 +25,6 @@ public class LottoInputView {
 
     public static int inputBonusNumber() {
         System.out.println("보너스 볼을 입력해 주세요.");
-        return sc.nextInt();
+        return Integer.parseInt(sc.nextLine());
     }
 }
